requirements.txt
streamlit
langchain
langchain_ollama
selenium
beautifulsoup4
lxml
html5lib
python-dotenv
webdriver-manager
langchain-google-genai 
google-generativeai
python-dotenv